This is a big mart sales prediction.
This is a sales prediction dataset that determine the amount of sales that was made. We had item fat content that contains product that are regular and product that has low fat. We had item type that had 16 unique items, item type represents the category which the product belongs to. Outlet type represent whether the outlet is just a grocery store or some sort of supermarket. In the dataset we discovered that we had more of items with low fat but we sold more of regular items. we also sold more items in the year 1985. we also had more items that are medium size. In the dataset, we had outliers in item visibility column but it was taken care of in the analysis. Finally the dataset was cleaned, we carried out EDA and we made predictions.